requirements.txt
amqp==2.2.2
anyjson==0.3.3
billiard==3.5.0.3
celery==4.1.0
Cython==0.26.1
falcon==1.3.0
hiredis==0.2.0
kombu==4.1.0
psycopg2==2.7.3.1
python-mimeparse==1.6.0
pytz==2017.2
PyYAML==3.12
redis==2.10.6
six==1.11.0
ujson==1.35
uWSGI==2.0.15
vine==1.1.4
